Refactor to abstraction for cleaner implementation
