+++ src/scan-app/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "Deep scan web app for the FOSS compliance of a npm package",
   "scripts": {
+    "serve": "bunx run --hot ./index.tsx"
   },
   "keywords": [],
   "license": "MIT",
