+++ src/scan-cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@senthanal/deep-scan-cli",
-  "version": "2.1.1",
+  "version": "2.2.0",
   "description": "Check a npm package FOSS compliance with a single command",
   "bin": {
     "deep-scan-cli": "./index.js"
