+++ src/scan-lib/index.ts
@@ -4,4 +4,5 @@ export * from './types';
 export * from './ScanLogger';
 export * from './UILogger';
 export * from './TerminalLogger';
-
+export * from './ScanProject';
+export * from './ScanPackage';
